Fix an issue with the contract archive dashboard
- It was reporting the contract of the location of the participant
- Non-contract areas were being added to the last column (making figures wrong)
