+++ src/Server.UI/Components/Dashboard/QA/ArchivedCasesByTenantAndReasonDashboardComponent.razor
@@ -1,0 +1,117 @@
+@using Cfo.Cats.Application.Features.Dashboard.Queries
+@inherits CatsComponent<GetArchivedCasesByTenantAndReason.ArchivedCasesByTenantAndReasonDto>
+
+@if (Loading)
+{
+    <MudLoading Text="Loading" />
+}
+
+@if (ErrorMessage is not null)
+{
+    <MudAlert Variant="Variant.Filled" Severity="Severity.Error">
+        @ErrorMessage
+    </MudAlert>
+}
+
+@if (Data is not null && Data.ChartData.Any())
+{
+
+    <MudAlert Severity="Severity.Info" Variant="Variant.Filled" Class="ma-4">
+        This data is based on the tenant of the user who did the archiving regardless of the contract location of the participant at the time of archive.
+    </MudAlert>
+
+    if (VisualMode)
+    {
+        <ApexCharts.ApexChart TItem="GetArchivedCasesByTenantAndReason.ArchivedCasesChartData"
+                              Options="@Options"
+                              Height="650"
+                              Width="@( "100%" )">
+
+            @foreach (var series in SeriesByReason)
+            {
+                <ApexCharts.ApexPointSeries
+                    TItem="GetArchivedCasesByTenantAndReason.ArchivedCasesChartData"
+                    Items="@(series.Items)"
+                    SeriesType="ApexCharts.SeriesType.Bar"
+                    Name="@(series.Reason)"
+                    XValue="@(e => e.Tenant)"
+                    YValue="@(e => e.Count)"
+                    ShowDataLabels="true" />
+            }
+
+        </ApexCharts.ApexChart>
+    }
+    else
+    {
+        <MudTable Items="@Data.TabularData" MultiSelection="false" Striped="true" Hover="true">
+            <HeaderContent>               
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.Tenant)">
+                        Tenant 
+                    </MudTableSortLabel>
+                </MudTh>
+
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.Reason)">
+                        Reason
+                    </MudTableSortLabel>
+                </MudTh>
+
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.ParticipantId)">
+                        Participant
+                    </MudTableSortLabel>
+                </MudTh>
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.Location)">
+                        Location (at archive)
+                    </MudTableSortLabel>
+                </MudTh>
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.Created)">
+                        Created
+                    </MudTableSortLabel>
+                </MudTh>
+
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.From)">
+                        From
+                    </MudTableSortLabel>
+                </MudTh>
+
+                <MudTh>
+                    <MudTableSortLabel SortBy="new Func<GetArchivedCasesByTenantAndReason.ArchivedCasesTabularData, object>(x => x.To.GetValueOrDefault())">
+                        To
+                    </MudTableSortLabel>
+                </MudTh>
+
+                <MudTh>
+                    Created By
+                </MudTh>
+            </HeaderContent>
+
+            <RowTemplate>
+                <MudTd DataLabel="Tenant">@context.Tenant</MudTd>
+                <MudTd DataLabel="Reason">@context.Reason</MudTd>
+                <MudTd DataLabel="Participant">@context.ParticipantId</MudTd>
+                <MudTd DataLabel="Location">@context.Location</MudTd>
+                <MudTd DataLabel="Created">@context.Created.ToShortDateString()</MudTd>
+                <MudTd DataLabel="From">@context.From.ToShortDateString()</MudTd>
+                <MudTd DataLabel="To">@context.To?.ToShortDateString()</MudTd>
+                <MudTd DataLabel="Created By">@context.CreatedBy</MudTd>
+            </RowTemplate>
+
+            <PagerContent>
+                <MudTablePager PageSizeOptions="new int[] { 10, 20 }" />
+            </PagerContent>
+        </MudTable>
+    }
+}
+else
+{
+    <MudText Typo="Typo.body2"
+             Color="Color.Secondary"
+             Align="MudBlazor.Align.Center">
+        No data available
+    </MudText>
+}
