+++ src/Server.UI/Pages/Dashboard/PerformanceDashboard.razor
@@ -42,7 +42,7 @@
 <MudStack>
     <MudTabs Class="pa-6" ApplyEffectsToContainer="true" Outlined="true">
         <MudTabPanel Text="Archived Case Behaviour">
-            <ArchivedCasesByContractAndReasonDashboardComponent TenantId="@SelectedTenantId" DateRange="@_dateRange" @key="SelectedTenantId + (_dateRange.Start?.Ticks ?? 0) + (_dateRange.End?.Ticks ?? 0)" VisualMode="@_visualMode"/>
+            <ArchivedCasesByTenantAndReasonDashboardComponent TenantId="@SelectedTenantId" DateRange="@_dateRange" @key="SelectedTenantId + (_dateRange.Start?.Ticks ?? 0) + (_dateRange.End?.Ticks ?? 0)" VisualMode="@_visualMode"/>
         </MudTabPanel>
     </MudTabs>
 </MudStack>
